requirements.txt
# requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
streamlit==1.28.0
google-generativeai==0.3.0
chromadb==0.4.15
pypdf2==3.0.1
python-docx==1.1.0
html2text==2020.1.16
numpy==1.24.3
pydantic==2.4.2
python-multipart==0.0.6
requests==2.31.0
aiofiles==23.2.1
python-jose==3.3.0
passlib==1.7.4
python-dotenv==1.0.0